requirements.txt
# AWS S3 Storage Integration - Dependencies
# Python 3.10+

# Core Django & DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# AWS S3 Storage
boto3>=1.34.0
django-storages>=1.14.0

# Celery for async tasks
celery>=5.3.0
redis>=5.0.0

# Environment variables
python-decouple>=3.8

# File processing & AI
Pillow>=10.0.0  # Image processing
PyPDF2>=3.0.0   # PDF handling
python-magic>=0.4.27  # File type detection

# Optional: AI/OCR Libraries (uncomment as needed)
# pytesseract>=0.3.10  # OCR
# opencv-python>=4.8.0  # Image processing for OCR

# PostgreSQL (already in your stack)
psycopg2-binary>=2.9.9

# Utilities
python-dateutil>=2.8.2

